Refactoring Carousel component code from HomePage, adding validation and destructuring
