+++ client/app/components/HomePageTemplate/index.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router';
 import Helmet from 'react-helmet';
-import Slider from 'react-slick';
 import Reveal from 'react-reveal';
 
 // Components
+import Carousel from './Carousel';
 import QuickFacts from './QuickFacts';
 import ServiceIcons from './ServiceIcons';
 import WhatWeDo from './WhatWeDo';
 
 // Styles
-import bulma from 'styles/bulma.scss';
+import bulma from 'styles/bulma.scss'; // eslint-disable-line import/first
 import styles from './styles.scss';
 
 const boldHero = `${bulma.hero} ${bulma['is-bold']}`;
@@ -19,47 +18,6 @@ const mainHero = `${bulma.hero} ${bulma['is-fullheight']} ${bulma['is-primary']}
 const infoHero = `${boldHero} ${bulma['is-medium']} ${bulma['is-info']}`;
 const lightHero = `${boldHero} ${bulma['is-medium']} ${bulma['is-light']}`;
 const mediumCustomContent = `${bulma.content} ${bulma['is-medium']} ${styles.content}`;
-
-const sliders = (data) => data.map((slider, idx) => (
-  <div className={`${bulma.container} ${styles.sliderContainer}`} key={idx}>
-    <div className={`${bulma.content} ${bulma['has-text-centered']}`}>
-      <h1>{slider.headline}</h1>
-      <h3>{slider.secondaryText}</h3>
-      <Link to={slider.link.route}>{slider.link.text}</Link>
-    </div>
-  </div>
-));
-
-const LeftNav = (props) =>
-  <button className={styles.leftNav} onClick={props.onClick}>
-    <span className={bulma.icon} aria-hidden="true">
-      <i className="fa fa-arrow-left fa-4x"></i>
-    </span>
-  </button>;
-
-const RightNav = (props) =>
-  <button className={styles.rightNav} onClick={props.onClick}>
-    <span className={bulma.icon} aria-hidden="true">
-      <i className="fa fa-arrow-right fa-4x"></i>
-    </span>
-  </button>;
-
-const sliderSettings = {
-  prevArrow: <LeftNav />,
-  nextArrow: <RightNav />,
-  dots: true,
-  autoplay: true,
-  autoplaySpeed: 4000,
-  fade: true,
-  pauseOnHover: true,
-  slidesToShow: 1,
-  slidesToScroll: 1,
-  draggable: false,
-  className: 'center',
-  centerMode: true,
-  infinite: true,
-  centerPadding: '100px',
-};
 
 const HomePageTemplate = ({ data }) => (
   <section>
@@ -87,9 +45,7 @@ const HomePageTemplate = ({ data }) => (
         <div className={bulma['hero-body']}>
           <div className={bulma.container}>
             <Reveal effect="animated fadeIn">
-              <Slider {...sliderSettings}>
-                {sliders(data.sliders)}
-              </Slider>
+              <Carousel slides={data.sliders} />
             </Reveal>
           </div>
         </div>
